Dodaje endpoint do dodawania produktów za pomocą CSV Fix przy usuwaniu

+++ src/main/java/pl/michalkarwowski/api/repositories/InvoicePositionRepository.java
@@ -6,5 +6,5 @@ import pl.michalkarwowski.api.models.InvoicePosition;
 import java.util.List;
 
 public interface InvoicePositionRepository extends CrudRepository<InvoicePosition, Long> {
-    List<InvoicePosition> findAllByProductId(Long id);
+    List<InvoicePosition> findAllByProductId(Integer id);
 }
